refactor class library to remove ClassObject
