+++ starfish/agent/__init__.py
@@ -1,7 +1,5 @@
 
 
-from starfish.agent.agent_object import AgentObject
+from starfish.agent.agent import Agent
 from starfish.agent.listing_agent import ListingAgent
 from starfish.agent.squid_agent import SquidAgent
-
-
