+++ starfish/agent/agent.py
@@ -3,7 +3,7 @@
     Basic agent object
 """
 
-class AgentObject():
+class Agent():
     def __init__(self, ocean):
         """init the the Ocean Object Base with the ocean instance"""
         self._ocean = ocean
