+++ starfish/metadata/__init__.py
@@ -1,4 +1,2 @@
 
-from starfish.metadata.metadata_object import MetadataObject
 from starfish.metadata.metadata import Metadata
-
